Refactor and decompose edit course structure elements
