+++ apps/platform/src/lib/infrastructure/client/pages/workspace/edit/components/lesson-item.tsx
@@ -1,0 +1,67 @@
+import {
+    Button,
+    CheckBox,
+    ContentControlButtons,
+    IconLesson,
+    InputField,
+} from '@maany_shr/e-class-ui-kit';
+import { CourseLesson } from '../types';
+
+interface LessonItemProps {
+    lesson: CourseLesson;
+    onMoveUp: () => void;
+    onMoveDown: () => void;
+    onDelete: () => void;
+    isFirst: boolean;
+    isLast: boolean;
+    onTitleChange: (newTitle: string) => void;
+    onExtraTrainingChange: (isExtraTraining: boolean) => void;
+}
+
+export function LessonItem({
+    lesson,
+    onMoveUp,
+    onMoveDown,
+    onDelete,
+    isFirst,
+    isLast,
+    onTitleChange,
+    onExtraTrainingChange,
+}: LessonItemProps) {
+    return (
+        <div className="flex flex-col gap-4 bg-card-fill border border-base-neutral-700 rounded-lg p-3">
+            <div className="flex gap-4 items-center w-full">
+                <IconLesson />
+                <InputField
+                    value={lesson.title || ''}
+                    inputText="Lesson title"
+                    setValue={(value) => {
+                        onTitleChange(value);
+                    }}
+                    className="flex-1 font-bold"
+                />
+                <ContentControlButtons
+                    onMoveUp={onMoveUp}
+                    onMoveDown={onMoveDown}
+                    onDelete={onDelete}
+                    isFirst={isFirst}
+                    isLast={isLast}
+                />
+            </div>
+            <CheckBox
+                name=""
+                label="Extra training"
+                value=""
+                checked={lesson.isExtraTraining}
+                onChange={() => onExtraTrainingChange(!lesson.isExtraTraining)}
+                withText
+                className="w-fit"
+            />
+            <Button
+                variant="secondary"
+                text={lesson.id ? 'Edit' : 'Save to edit'}
+                disabled={lesson.id === undefined}
+            />
+        </div>
+    );
+}
