+++ apps/platform/src/lib/infrastructure/client/pages/workspace/edit/edit-course-content.tsx
@@ -1,392 +1,25 @@
 import {
-    Button,
-    CheckBox,
-    IconChevronDown,
-    IconChevronUp,
+    ComponentCard,
     IconLesson,
     IconMilestone,
-    IconMinus,
     IconModule,
-    IconPlus,
-    IconTrashAlt,
-    InputField,
 } from '@maany_shr/e-class-ui-kit';
 import { TLocale } from '@maany_shr/e-class-translations';
-import {
-    DefaultError,
-    DefaultLoading,
-    Divider,
-} from '@maany_shr/e-class-ui-kit';
+import { DefaultError, DefaultLoading } from '@maany_shr/e-class-ui-kit';
 import { useLocale } from 'next-intl';
 import { trpc } from '../../../trpc/client';
 import { viewModels } from '@maany_shr/e-class-models';
 import { useGetCourseStructurePresenter } from '../../../hooks/use-course-structure-presenter';
 import { useEffect, useState } from 'react';
+import { ModuleEditor } from './components/module-editor';
+import {
+    ContentType,
+    CourseLesson,
+    CourseMilestone,
+    CourseModule,
+    EditCourseContentProps,
+} from './types';
 
-interface EditCourseContentProps {
-    slug: string;
-}
-
-interface StructureComponentProps {
-    name: string;
-    icon: React.ReactNode;
-    onClick?: () => void;
-}
-
-function StructureComponent({ name, icon, onClick }: StructureComponentProps) {
-    return (
-        <div
-            className="flex items-center gap-2 bg-base-neutral-800 border border-base-neutral-700 rounded-lg p-4 cursor-pointer"
-            onClick={onClick}
-        >
-            {icon}
-            <span className="text-lg">{name}</span>
-        </div>
-    );
-}
-
-enum ContentType {
-    Lesson = 'lesson',
-    Milestone = 'milestone',
-}
-
-interface CourseLesson {
-    type: ContentType.Lesson;
-    id?: number;
-    title?: string;
-    isExtraTraining: boolean;
-}
-
-interface CourseMilestone {
-    type: ContentType.Milestone;
-    id?: number;
-}
-
-interface CourseModule {
-    id?: number;
-    title?: string;
-    content: (CourseLesson | CourseMilestone)[];
-}
-
-interface ModuleContentProps {
-    content: (CourseLesson | CourseMilestone)[];
-    onMoveContentUp: (contentIndex: number) => void;
-    onMoveContentDown: (contentIndex: number) => void;
-    onDeleteContent: (contentIndex: number) => void;
-    onLessonTitleChange: (lessonIndex: number, newTitle: string) => void;
-    onLessonExtraTrainingChange: (
-        lessonIndex: number,
-        extraTraining: boolean,
-    ) => void;
-}
-
-function ContentControlButtons({
-    onMoveUp,
-    onMoveDown,
-    onDelete,
-    isExpanded = false,
-    onExpand,
-    isFirst,
-    isLast,
-}: {
-    onMoveUp: () => void;
-    onMoveDown: () => void;
-    onDelete: () => void;
-    isExpanded?: boolean;
-    onExpand?: () => void;
-    isFirst: boolean;
-    isLast: boolean;
-}) {
-    return (
-        <div className="flex gap-2">
-            <div className="flex items-center gap-2">
-                <Button
-                    variant="text"
-                    iconLeft={<IconTrashAlt />}
-                    hasIconLeft
-                    onClick={() => onDelete()}
-                    className="px-0"
-                />
-                <Button
-                    variant="text"
-                    iconLeft={<IconChevronUp />}
-                    hasIconLeft
-                    onClick={() => onMoveUp()}
-                    disabled={isFirst}
-                    className="px-0"
-                />
-                <Button
-                    variant="text"
-                    iconLeft={<IconChevronDown />}
-                    hasIconLeft
-                    onClick={() => onMoveDown()}
-                    disabled={isLast}
-                    className="px-0"
-                />
-                {onExpand && (
-                    <Button
-                        variant="text"
-                        iconLeft={isExpanded ? <IconMinus /> : <IconPlus />}
-                        hasIconLeft
-                        onClick={() => onExpand()}
-                        className="px-0"
-                    />
-                )}
-            </div>
-        </div>
-    );
-}
-
-function LessonItem({
-    lesson,
-    onMoveUp,
-    onMoveDown,
-    onDelete,
-    isFirst,
-    isLast,
-    onTitleChange,
-    onExtraTrainingChange,
-}: {
-    lesson: CourseLesson;
-    onMoveUp: () => void;
-    onMoveDown: () => void;
-    onDelete: () => void;
-    isFirst: boolean;
-    isLast: boolean;
-    onTitleChange: (newTitle: string) => void;
-    onExtraTrainingChange: (isExtraTraining: boolean) => void;
-}) {
-    return (
-        <div className="flex flex-col gap-4 bg-card-fill border border-base-neutral-700 rounded-lg p-3">
-            <div className="flex gap-4 items-center w-full">
-                <IconLesson />
-                <InputField
-                    value={lesson.title || ''}
-                    inputText="Lesson title"
-                    setValue={(value) => {
-                        onTitleChange(value);
-                    }}
-                    className="flex-1 font-bold"
-                />
-                <ContentControlButtons
-                    onMoveUp={onMoveUp}
-                    onMoveDown={onMoveDown}
-                    onDelete={onDelete}
-                    isFirst={isFirst}
-                    isLast={isLast}
-                />
-            </div>
-            <CheckBox
-                name=""
-                label="Extra training"
-                value=""
-                checked={lesson.isExtraTraining}
-                onChange={() => onExtraTrainingChange(!lesson.isExtraTraining)}
-                withText
-                className="w-fit"
-            />
-            <Button
-                variant="secondary"
-                text={lesson.id ? 'Edit' : 'Save to edit'}
-                disabled={lesson.id === undefined}
-            />
-        </div>
-    );
-}
-
-function MilestoneItem({
-    milestone,
-    onMoveUp,
-    onMoveDown,
-    onDelete,
-    isFirst,
-    isLast,
-}: {
-    milestone: CourseMilestone;
-    onMoveUp: () => void;
-    onMoveDown: () => void;
-    onDelete: () => void;
-    isFirst: boolean;
-    isLast: boolean;
-}) {
-    return (
-        <div className="flex gap-4 items-center bg-card-fill border border-base-neutral-700 rounded-lg p-3">
-            <IconMilestone />
-            <span className="font-bold w-full">Milestone</span>
-            <ContentControlButtons
-                onMoveUp={onMoveUp}
-                onMoveDown={onMoveDown}
-                onDelete={onDelete}
-                isFirst={isFirst}
-                isLast={isLast}
-            />
-        </div>
-    );
-}
-
-function ModuleContent({
-    content,
-    onMoveContentUp,
-    onMoveContentDown,
-    onDeleteContent,
-    onLessonTitleChange,
-    onLessonExtraTrainingChange,
-}: ModuleContentProps) {
-    const isEmpty = content.length === 0;
-
-    return (
-        <div className="flex flex-col gap-2 ml-4">
-            {isEmpty && (
-                <div className=" flex flex-col gap-2 bg-card-fill border border-base-neutral-700 rounded-lg p-4">
-                    <span className="text-text-secondary">
-                        Add lessons or milestones
-                    </span>
-                </div>
-            )}
-            {content.map((item, index) => {
-                if (item.type === ContentType.Milestone) {
-                    return (
-                        <MilestoneItem
-                            key={`milestone-${index}`}
-                            milestone={item}
-                            onMoveUp={() => onMoveContentUp(index)}
-                            onMoveDown={() => onMoveContentDown(index)}
-                            onDelete={() => onDeleteContent(index)}
-                            isFirst={index === 0}
-                            isLast={index === content.length - 1}
-                        />
-                    );
-                }
-                if (item.type === ContentType.Lesson) {
-                    return (
-                        <LessonItem
-                            key={`lesson-${index}`}
-                            lesson={item}
-                            onMoveUp={() => onMoveContentUp(index)}
-                            onMoveDown={() => onMoveContentDown(index)}
-                            onDelete={() => onDeleteContent(index)}
-                            isFirst={index === 0}
-                            isLast={index === content.length - 1}
-                            onTitleChange={(newTitle) =>
-                                onLessonTitleChange(index, newTitle)
-                            }
-                            onExtraTrainingChange={(isExtraTraining) =>
-                                onLessonExtraTrainingChange(
-                                    index,
-                                    isExtraTraining,
-                                )
-                            }
-                        />
-                    );
-                }
-                return null;
-            })}
-        </div>
-    );
-}
-
-interface ModuleEditorProps {
-    module: CourseModule;
-    onUpdate: (updatedModule: CourseModule) => void;
-    onDelete: () => void;
-    onMoveUp: () => void;
-    onMoveDown: () => void;
-    onMoveContentUp: (contentIndex: number) => void;
-    onMoveContentDown: (contentIndex: number) => void;
-    onDeleteContent: (contentIndex: number) => void;
-    isExpanded: boolean;
-    onExpand: () => void;
-    isFirst: boolean;
-    isLast: boolean;
-
-    onLessonTitleChange: (index: number, title: string) => void;
-    onLessonExtraTrainingChange: (
-        index: number,
-        isExtraTraining: boolean,
-    ) => void;
-}
-
-export function ModuleEditor({
-    module,
-    onUpdate,
-    onDelete,
-    onMoveUp,
-    onMoveDown,
-    onMoveContentUp,
-    onMoveContentDown,
-    onDeleteContent,
-    isExpanded,
-    onExpand,
-    isFirst,
-    isLast,
-    onLessonTitleChange,
-    onLessonExtraTrainingChange,
-}: ModuleEditorProps) {
-    const handleTitleChange = (value: string) => {
-        onUpdate({ ...module, title: value });
-    };
-
-    return (
-        <div className="flex flex-col gap-2">
-            <div className="flex items-center justify-between gap-4 bg-card-fill border border-base-neutral-700 rounded-lg p-4">
-                <div className="flex items-center gap-4 flex-1">
-                    <div className="h-12 w-12 flex-shrink-0 bg-base-neutral-700 border border-base-neutral-600 rounded-lg flex items-center justify-center">
-                        <IconModule />
-                    </div>
-                    <InputField
-                        value={module.title || ''}
-                        inputText="Module title"
-                        setValue={handleTitleChange}
-                        className="w-full font-bold"
-                    />
-                </div>
-                <ContentControlButtons
-                    onMoveUp={onMoveUp}
-                    onMoveDown={onMoveDown}
-                    onDelete={onDelete}
-                    isFirst={isFirst}
-                    isLast={isLast}
-                    isExpanded={isExpanded}
-                    onExpand={onExpand}
-                />
-            </div>
-            {isExpanded && (
-                <ModuleContent
-                    content={module.content}
-                    onMoveContentUp={(contentIndex) =>
-                        onMoveContentUp(contentIndex)
-                    }
-                    onMoveContentDown={(contentIndex) =>
-                        onMoveContentDown(contentIndex)
-                    }
-                    onDeleteContent={(contentIndex) =>
-                        onDeleteContent(contentIndex)
-                    }
-                    onLessonTitleChange={(lessonIndex, newTitle) =>
-                        onLessonTitleChange(lessonIndex, newTitle)
-                    }
-                    onLessonExtraTrainingChange={(
-                        lessonIndex,
-                        isExtraTraining,
-                    ) =>
-                        onLessonExtraTrainingChange(
-                            lessonIndex,
-                            isExtraTraining,
-                        )
-                    }
-                />
-            )}
-        </div>
-    );
-}
-
-// TODO: Translate
-// TODO: Add search functionality
-// TODO: Add lesson templating
-// TODO: Add limits on number of modules, lessons, and milestones
-// TODO: Integrate undo/redo functionality
-// TODO: Add verification on deletion of modules, lessons, and milestones
-// TODO: Scroll to newly created module, lesson, or milestone
 export default function EditCourseContent({ slug }: EditCourseContentProps) {
     const locale = useLocale() as TLocale;
 
@@ -707,17 +340,17 @@ export default function EditCourseContent({ slug }: EditCourseContentProps) {
             <div className="h-fit flex flex-col gap-3 bg-card-fill border border-base-neutral-700 rounded-lg p-4 lg:w-[300px] w-full">
                 <span className="text-lg font-bold">Components</span>
                 <div className="flex flex-col gap-2">
-                    <StructureComponent
+                    <ComponentCard
                         name="Module"
                         icon={<IconModule />}
                         onClick={() => addModule()}
                     />
-                    <StructureComponent
+                    <ComponentCard
                         name="Lesson"
                         icon={<IconLesson />}
                         onClick={() => addLesson()}
                     />
-                    <StructureComponent
+                    <ComponentCard
                         name="Milestone"
                         icon={<IconMilestone />}
                         onClick={() => addMilestone()}
